+++ module/icp/algs/blake3/blake3_generic.c
@@ -187,7 +187,8 @@ static inline void blake3_hash_many_generic(const uint8_t * const *inputs,
 	}
 }
 
-static inline boolean_t blake3_is_generic_supported(void)
+/* the generic implementation is always okay */
+static boolean_t blake3_is_supported(void)
 {
 	return (B_TRUE);
 }
@@ -196,7 +197,7 @@ const blake3_ops_t blake3_generic_impl = {
 	.compress_in_place = blake3_compress_in_place_generic,
 	.compress_xof = blake3_compress_xof_generic,
 	.hash_many = blake3_hash_many_generic,
-	.is_supported = blake3_is_generic_supported,
+	.is_supported = blake3_is_supported,
 	.degree = 4,
 	.name = "generic"
 };
